+++ spirv_reflect.c
@@ -52,11 +52,11 @@ enum {
 };
 
 enum {
-  INVALID_VALUE  = 0xFFFFFFFF,
+  INVALID_VALUE = (int)0xFFFFFFFF,
 };
 
 enum {
-  MAX_NODE_NAME_LENGTH        = 1024,
+  MAX_NODE_NAME_LENGTH                 = 1024,
   // Number of unique PhysicalStorageBuffer structs tracked to detect recursion
   MAX_RECURSIVE_PHYSICAL_POINTER_CHECK = 128,
 };
@@ -1345,7 +1345,7 @@ static SpvReflectResult ParseMemberCounts(SpvReflectPrvParser* p_parser) {
         continue;
       }
 
-      if (member_index == INVALID_VALUE) {
+      if (member_index == (uint32_t)INVALID_VALUE) {
         return SPV_REFLECT_RESULT_ERROR_RANGE_EXCEEDED;
       }
 
@@ -1789,7 +1789,7 @@ static SpvReflectResult ParseType(SpvReflectPrvParser* p_parser, SpvReflectPrvNo
   if (result == SPV_REFLECT_RESULT_SUCCESS) {
     // Since the parse descends on type information, these will get overwritten
     // if not guarded against assignment. Only assign if the id is invalid.
-    if (p_type->id == INVALID_VALUE) {
+    if (p_type->id == (uint32_t)INVALID_VALUE) {
       p_type->id = p_node->result_id;
       p_type->op = p_node->op;
       p_type->decoration_flags = 0;
@@ -2144,7 +2144,7 @@ static SpvReflectResult ParseDescriptorBindings(SpvReflectPrvParser* p_parser, S
          (p_node->storage_class != SpvStorageClassUniformConstant))) {
       continue;
     }
-    if ((p_node->decorations.set.value == INVALID_VALUE) || (p_node->decorations.binding.value == INVALID_VALUE)) {
+    if ((p_node->decorations.set.value == (uint32_t)INVALID_VALUE) || (p_node->decorations.binding.value == (uint32_t)INVALID_VALUE)) {
       continue;
     }
 
@@ -2179,7 +2179,7 @@ static SpvReflectResult ParseDescriptorBindings(SpvReflectPrvParser* p_parser, S
          (p_node->storage_class != SpvStorageClassUniformConstant))) {
       continue;
     }
-    if ((p_node->decorations.set.value == INVALID_VALUE) || (p_node->decorations.binding.value == INVALID_VALUE)) {
+    if ((p_node->decorations.set.value == (uint32_t)INVALID_VALUE) || (p_node->decorations.binding.value == (uint32_t)INVALID_VALUE)) {
       continue;
     }
 
@@ -3446,7 +3446,7 @@ static SpvReflectResult ParseByteAddressBuffer(SpvReflectPrvParser* p_parser, Sp
     // All arithmetic ops takes 2 operands, assumption is the 2nd operand has the constant
     UNCHECKED_READU32(p_parser, p_next_node->word_offset + 4, base_id);
     uint32_t value = GetUint32Constant(p_parser, base_id);
-    if (value == INVALID_VALUE) {
+    if (value == (uint32_t)INVALID_VALUE) {
       return not_found;
     }
 
@@ -5022,7 +5022,7 @@ const SpvReflectDescriptorSet* spvReflectGetEntryPointDescriptorSet(const SpvRef
 
 const SpvReflectInterfaceVariable* spvReflectGetInputVariableByLocation(const SpvReflectShaderModule* p_module, uint32_t location,
                                                                         SpvReflectResult* p_result) {
-  if (location == INVALID_VALUE) {
+  if (location == (uint32_t)INVALID_VALUE) {
     if (IsNotNull(p_result)) {
       *p_result = SPV_REFLECT_RESULT_ERROR_ELEMENT_NOT_FOUND;
     }
@@ -5052,7 +5052,7 @@ const SpvReflectInterfaceVariable* spvReflectGetInputVariable(const SpvReflectSh
 const SpvReflectInterfaceVariable* spvReflectGetEntryPointInputVariableByLocation(const SpvReflectShaderModule* p_module,
                                                                                   const char* entry_point, uint32_t location,
                                                                                   SpvReflectResult* p_result) {
-  if (location == INVALID_VALUE) {
+  if (location == (uint32_t)INVALID_VALUE) {
     if (IsNotNull(p_result)) {
       *p_result = SPV_REFLECT_RESULT_ERROR_ELEMENT_NOT_FOUND;
     }
@@ -5155,7 +5155,7 @@ const SpvReflectInterfaceVariable* spvReflectGetEntryPointInputVariableBySemanti
 
 const SpvReflectInterfaceVariable* spvReflectGetOutputVariableByLocation(const SpvReflectShaderModule* p_module, uint32_t location,
                                                                          SpvReflectResult* p_result) {
-  if (location == INVALID_VALUE) {
+  if (location == (uint32_t)INVALID_VALUE) {
     if (IsNotNull(p_result)) {
       *p_result = SPV_REFLECT_RESULT_ERROR_ELEMENT_NOT_FOUND;
     }
@@ -5185,7 +5185,7 @@ const SpvReflectInterfaceVariable* spvReflectGetOutputVariable(const SpvReflectS
 const SpvReflectInterfaceVariable* spvReflectGetEntryPointOutputVariableByLocation(const SpvReflectShaderModule* p_module,
                                                                                    const char* entry_point, uint32_t location,
                                                                                    SpvReflectResult* p_result) {
-  if (location == INVALID_VALUE) {
+  if (location == (uint32_t)INVALID_VALUE) {
     if (IsNotNull(p_result)) {
       *p_result = SPV_REFLECT_RESULT_ERROR_ELEMENT_NOT_FOUND;
     }
